moved search from srx/app/api/search/route to src/app/(frontend)/search
