+++ playwright.config.ts
@@ -39,7 +39,7 @@ export default defineConfig({
     },
   ],
   webServer: {
-    command: process.env.CI ? 'pnpm start' : 'pnpm dev',
+    command: 'pnpm dev',
     reuseExistingServer: true,
     url: 'http://127.0.0.1:3000',
     /* timeout 120 seconds (120000ms) */
